Create Triangle.java
Write a program to validate the side of the triangle.  Get three sides measurement from user and validate the triangle. 

Sample Input:

Enter the value for side1

10

Enter the value for side2

10

Enter the value for side3

10

Sample output-1


Sides form a triangle

Sample Input-2

Enter the value for side1

1

Enter the value for side2

2

Enter the value for side3

3